Add JRuby support
JRuby uses jdbc-sqlite3 rather than sqlite3 and the connection string needs to be prefixed with jdbc:. Explicitly requiring sqlite3 is not necessary as Sequel does this automatically. JRuby doesn't have bundler ~> 1.0.0 and putting bundler in your Gemfile is redundant anyway. mysql has also been removed from the Gemfile because it is only used within the api directory, which has its own Gemfile. The presence of a gemspec may make the rest of this Gemfile redundant too but that it outside the scope of supporting JRuby.